+++ frontend/src/App.jsx
@@ -75,6 +75,7 @@ import DipendentiAnagrafica from "./pages/dipendenti/DipendentiAnagrafica";
 import DipendentiTurni from "./pages/dipendenti/DipendentiTurni";
 import FoglioSettimana from "./pages/dipendenti/FoglioSettimana";
 import VistaMensile from "./pages/dipendenti/VistaMensile";
+import PerDipendente from "./pages/dipendenti/PerDipendente";
 import GestioneReparti from "./pages/dipendenti/GestioneReparti";
 import DipendentiCosti from "./pages/dipendenti/DipendentiCosti";
 import DipendentiBustePaga from "./pages/dipendenti/DipendentiBustePaga";
@@ -260,6 +261,7 @@ export default function App() {
         <Route path="/dipendenti/anagrafica" element={<ProtectedRoute module="dipendenti"><DipendentiAnagrafica /></ProtectedRoute>} />
         <Route path="/dipendenti/turni" element={<ProtectedRoute module="dipendenti"><FoglioSettimana /></ProtectedRoute>} />
         <Route path="/dipendenti/turni/mese" element={<ProtectedRoute module="dipendenti"><VistaMensile /></ProtectedRoute>} />
+        <Route path="/dipendenti/turni/dipendente" element={<ProtectedRoute module="dipendenti"><PerDipendente /></ProtectedRoute>} />
         <Route path="/dipendenti/turni-legacy" element={<ProtectedRoute module="dipendenti"><DipendentiTurni /></ProtectedRoute>} />
         <Route path="/dipendenti/reparti" element={<ProtectedRoute module="dipendenti"><GestioneReparti /></ProtectedRoute>} />
         <Route path="/dipendenti/costi" element={<ProtectedRoute module="dipendenti"><DipendentiCosti /></ProtectedRoute>} />